Reorganize pair converter code
Moved graph/optimization code from abstract_pair_converter_plugin.py to abstract_ccxt_input_plugin.py. This is because:
- the code was interfering with simple plugins like Historic-Crypto, which was printing the following message: "No Fiat Access Key. Fiat pricing will NOT be available. To enable fiat pricing, an access key from exchangerate.host is required.". This was confusing because Historic-Crypto has no dependency on exchangerate.host.
- it's cleaner to have a bare bone top class (abstract_pair_converter_plugin.py) and a graph/optimization-aware subclass (abstract_ccxt_input_plugin.py). This way, simple plugins don't have to inherit the extra complexity of optimization.
